requirements-deb.txt
python-daemon
python-dns
